requirements.txt
absl-py==1.4.0
accelerate==0.23.0
aiofiles==23.2.1
aiohttp==3.9.1
aioprocessing==2.0.1
aiosignal==1.3.1
analytics-python==1.2.9
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
biopython==1.81
black==23.10.0
bleach==6.1.0
blinker==1.7.0
braceexpand==0.1.7
bs4==0.0.1
captum==0.6.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.6.3
datasets==2.14.4
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
distro==1.8.0
docker-pycreds==0.4.0
einops==0.6.1
et-xmlfile==1.1.0
evaluate==0.4.0
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.104.1
fastjsonschema==2.18.1
filelock==3.12.3
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.1
gitdb==4.0.11
GitPython==3.1.40
gql==3.4.1
graphql-core==3.2.3
greenlet==3.0.2
h11==0.14.0
h5==0.8.10
h5py==3.9.0
httpcore==1.0.1
httpx==0.25.1
huggingface-hub==0.16.4
hydra-core==1.3.2
idna==3.4
ipykernel==6.25.1
ipynbname==2023.2.0.0
ipython==8.15.0
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
janus==1.0.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonlines==4.0.0
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.1
kiwisolver==1.4.5
lightning-utilities==0.9.0
llama-index==0.9.15.post2
llvmlite==0.41.1
loguru==0.7.2
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mauve-text==0.3.0
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy==1.6.1
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.1
objgraph==3.6.0
omegaconf==2.3.0
openai==1.2.2
opencv-python==4.8.0.76
openpyxl==3.1.2
outcome==1.3.0
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandas-stubs==2.1.1.230928
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.10.0
plotly==5.18.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydicom==2.4.3
Pygments==2.16.1
pynndescent==0.5.10
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.8.8
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rouge-score==0.1.2
rpds-py==0.12.0
safetensors==0.3.3
schedule==1.2.1
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.13.0
Send2Trash==1.8.2
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.2
starlette==0.27.0
supervisor==4.2.5
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.2.0
tiktoken==0.5.1
timm==0.9.12
tinycss2==1.2.1
tokenizers==0.13.3
torch==2.0.1
torchmetrics==1.2.0
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.33.0
trio==0.22.2
trio-websocket==0.11.1
typeguard==4.1.5
types-beautifulsoup4==4.12.0.6
types-html5lib==1.1.11.15
types-protobuf==4.24.0.4
types-python-dateutil==2.8.19.14
types-pytz==2023.3.1.1
types-tqdm==4.66.0.2
types-ujson==5.8.0.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
umap-learn==0.5.4
uri-template==1.3.0
wandb==0.16.0
wcwidth==0.2.6
weave==0.30.0
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.3.0
yarl==1.9.2